Remove 'web syncmedia'.
The functionality previously provided as part of 'web syncmedia' is now
provided by the Django 1.3 static file infrastructure.  Specifically,
'manage.py collectfiles'.  Furthermore this is now embedded as part of
'web start' which removes the requirement for a seperate command and as
such 'web syncmedia' is now deprecated, a no-op and prints out such a
message to STDOUT upon execution.


Former-commit-id: d76b387b856540bfbfa0245a30a9ffcc1107f772